VishGuard/voice_fetch/iphone_doggy.py
import os
import logging
from datetime import datetime
import sqlite3
from pathlib import Path
import json
from dotenv import load_dotenv

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class IPhoneVoiceFetcher:
    """Fetches voice call data from iPhone using local database."""
    
    def __init__(self):
        """Initialize iPhone voice fetcher."""
        load_dotenv()
        self.phone_number = os.getenv('IPHONE_NUMBER')
        self.db_path = self._get_call_db_path()
        
    def _get_call_db_path(self):
        """Get the path to iPhone call database."""
        # This would be the path to your iPhone backup
        backup_path = os.getenv('IPHONE_BACKUP_PATH')
        if not backup_path:
            raise ValueError("IPHONE_BACKUP_PATH not set in environment variables")
        return Path(backup_path) / '2b' / '2b2b0084a1bc3a5ac8c27afdf14afb42c61a19ca'
    
    def fetch_calls(self):
        """Fetch call records from iPhone database."""
        try:
            conn = sqlite3.connect(self.db_path)
            cursor = conn.cursor()
            
            # Query recent calls
            cursor.execute("""
                SELECT 
                    call.date,
                    call.duration,
                    call.phone_number,
                    call.call_type,
                    call.face_time_data
                FROM call
                WHERE call.date > strftime('%s', 'now', '-1 day')
                ORDER BY call.date DESC
            """)
            
            calls = []
            for row in cursor.fetchall():
                date, duration, number, call_type, face_time = row
                calls.append({
                    'date': datetime.fromtimestamp(date + 978307200).isoformat(),  # Apple's epoch
                    'duration': duration,
                    'caller_id': number,
                    'type': 'voice',
                    'call_type': call_type,
                    'is_facetime': bool(face_time)
                })
            
            conn.close()
            logger.info(f"Fetched {len(calls)} call records")
            return calls
            
        except Exception as e:
            logger.error(f"Error fetching calls: {e}")
            raise

    def monitor_realtime(self, callback):
        """Monitor calls in real-time using CallKit."""
        try:
            # This would use Apple's CallKit framework
            # to receive real-time call notifications
            pass
        except Exception as e:
            logger.error(f"Error in real-time monitoring: {e}")
            raise 