requirements.txt
# Core dependencies
python-dotenv==1.0.0
PyYAML==6.0.1
cryptography==41.0.7
watchdog==3.0.0

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.6.4

# ML and data processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Monitoring
prometheus-client==0.19.0
grafana-api-client==2.1.0

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Added from the code block
flask==2.0.1
rich==10.12.0
sentence-transformers==2.2.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
jinja2>=3.1.0
aiohttp>=3.9.0
markdown>=3.5.0
weasyprint>=60.1
pygments>=2.17.0

# Twilio integration
twilio>=8.10.0

# Email processing
imaplib2>=3.6
email-validator>=2.1.0

# Security and sanitization
bleach>=6.1.0
structlog>=24.1.0

# Testing
pytest-mock>=3.12.0

# Development tools
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0